move setup aggregator
